api/requirements.txt
Flask
gunicorn
setuptools
pymysql
firebase-admin
google-cloud-firestore
google-cloud-logging
SQLAlchemy>=1.3.15